Use QSFPM::setFilterRegularExpression everywhere
Instead of setFilterFixedString() which doesn't modify the QRegularExpression
part of the QSFPM, and use QRegularExpression::escape() on the pattern to make
it plain text.

Match case-insensitively to match what's done in
SortFilterModel::setFilterRegExp().

Note that this whole class is going to be moved out of here (see comment in
header file).